rename: MenuType, Operator 패키지 이동
